Create Attributes
The add_goal_point() method is responsible for adding a goal point for the player to reach in the game.

First, the code specifies the image path for the gate image and is constructed dynamically using a string and assumes that the gate image is located in the "img" directory and named "gate.png". 
The code then loads the gate image using pygame.image.load(img_path).

Next, the code scales the loaded image to match the size of a single grid cell (tile) in the game. This is achieved using pygame.transform.scale, which takes the loaded image and the desired dimensions as arguments. 
In this case, the dimensions are specified as (self.tile, self.tile), indicating that the image should be scaled to have a width and height equal to self.tile. Adjust this scaling factor as needed to match the desired size of the gate image on the screen.

Finally, the scaled gate image is drawn on the screen using screen.blit(img, (self.goal_cell.x * self.tile, self.goal_cell.y * self.tile)). The blit() function in Pygame is used to draw one Surface onto another. In this case, it takes the gate image img and specifies the position where it should be drawn on the screen. 
The position is determined by the x and y coordinates of the goal_cell (representing the grid cell where the goal point is located) multiplied by self.tile, which gives the pixel coordinates corresponding to the top-left corner of the grid cell.

The message() function only returns the winning message once was called.The is_game_over() method is responsible for checking if the player has reached the goal point in the game. 

First, the code calculates the absolute pixel coordinates of the goal cell by multiplying the x and y coordinates of the goal_cell with self.tile. This determines the top-left corner of the goal cell on the screen. Next, the method compares the player's current x and y coordinates with the goal cell's absolute coordinates. If both the player's x coordinate is greater than or equal to the goal cell's absolute x coordinate, and the player's y coordinate is greater than or equal to the goal cell's absolute y coordinate, it means that the player has reached or surpassed the goal cell. And if the player has reached the goal point, the method returns True, indicating that the game is over. Otherwise, it returns False, indicating that the game is still ongoing.